pyproject.toml
[project]
name = "kaleidoscope-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
  "markitdown[pdf]>=0.1.4",
  "pdfplumber>=0.11.9",
  "pymupdf>=1.26.7",
  "pytest>=9.0.2",
  "jinja2 (>=3.1.6,<4.0.0)",
  "pre-commit>=4.5.1",
  "pendulum>=3.2.0",
  "pydantic>=2.12.5",
  "loguru>=0.7.3",
  "dotenv>=0.9.9",
  "pyyaml>=6.0.3"
]

[tool.pytest.ini_options]
testpaths = [
  "tests"
]
addopts = "-ra -q -s"
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
minversion = "9.0"
markers = [
  "integration: marks tests as integration tests (require API key and --model)"
]
